Changed app to a single page with log
Changed app to a single logging page with all bluetooth connection logic statically defined. Fixed background bluetooth connection permissions. Fixed automatic bluetooth connection retry logic. Improved logging.
