+++ app/build.gradle
@@ -23,7 +23,7 @@ android {
     buildToolsVersion "29.0.0"
     defaultConfig {
         applicationId "com.punchthrough.blestarterappandroid"
-        minSdkVersion 26
+        minSdkVersion 29
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
