--- app/src/main/java/com/punchthrough/blestarterappandroid/ScanResultAdapter.kt
@@ -1,62 +1,0 @@
-/*
- * Copyright 2019 Punch Through Design LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.punchthrough.blestarterappandroid
-
-import android.bluetooth.le.ScanResult
-import android.view.View
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.synthetic.main.row_scan_result.view.device_name
-import kotlinx.android.synthetic.main.row_scan_result.view.mac_address
-import kotlinx.android.synthetic.main.row_scan_result.view.signal_strength
-import org.jetbrains.anko.layoutInflater
-
-class ScanResultAdapter(
-    private val items: List<ScanResult>,
-    private val onClickListener: ((device: ScanResult) -> Unit)
-) : RecyclerView.Adapter<ScanResultAdapter.ViewHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val view = parent.context.layoutInflater.inflate(
-            R.layout.row_scan_result,
-            parent,
-            false
-        )
-        return ViewHolder(view, onClickListener)
-    }
-
-    override fun getItemCount() = items.size
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val item = items[position]
-        holder.bind(item)
-    }
-
-    class ViewHolder(
-        private val view: View,
-        private val onClickListener: ((device: ScanResult) -> Unit)
-    ) : RecyclerView.ViewHolder(view) {
-
-        fun bind(result: ScanResult) {
-            view.device_name.text = result.device.name ?: "Unnamed"
-            view.mac_address.text = result.device.address
-            view.signal_strength.text = "${result.rssi} dBm"
-            view.setOnClickListener { onClickListener.invoke(result) }
-            onClickListener.invoke(result)
-        }
-    }
-}
