+++ app/src/main/res/layout/activity_ble_operations.xml
@@ -24,61 +24,13 @@
     tools:context=".BleOperationsActivity">
 
     <LinearLayout
-        android:id="@+id/mtu_container"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:orientation="horizontal"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content">
-
-        <EditText
-            android:id="@+id/mtu_field"
-            android:hint="MTU value"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="8dp"
-            android:importantForAutofill="no"
-            android:inputType="number"
-            android:digits="01234567890" />
-
-        <Button
-            android:id="@+id/request_mtu_button"
-            android:text="Request MTU"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
-
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/characteristics_title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:layout_marginStart="16dp"
-        android:text="Characteristics"
-        android:textStyle="bold"
-        android:textSize="16sp"
-        app:layout_constraintTop_toBottomOf="@id/mtu_container"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <LinearLayout
-        app:layout_constraintTop_toBottomOf="@id/characteristics_title"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         android:orientation="vertical"
         android:layout_width="match_parent"
         android:layout_height="0dp">
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/characteristics_recycler_view"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1"
-            android:scrollbarFadeDuration="0"
-            android:scrollbars="vertical"
-            tools:listitem="@layout/row_characteristic" />
 
         <TextView
             android:id="@+id/log_title"
